pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfy-meta"
version = "0.1.0"
description = "Extract EXIF and ComfyUI metadata from images"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0.0",
  "piexif>=1.1.3",
]

[project.scripts]
comfy-meta = "extractor.cli:main"
